feat: rewrite caching, refactor notifs, add new notif channels, tweak embeds
Rewrote caching so that instead of multiple random files, its 1 file
that holds everything from cached secrets, to shard dependent
configurations.

Refactored Notifs to be less of a pain to use, including method stacking
and similar.

Added new notif channel choices, such as `SubmitAd`, `SubmitUser`,
`VerifyAd`, `VerifyUser`.

Tweaked a couple embeds to be a bit prettier, though I am not sure if
everything is correct with the embeds, we'll find out